+++ src/main/java/dinkplugin/DinkPluginConfig.java
@@ -434,10 +434,21 @@ public interface DinkPluginConfig extends Config {
     }
 
     @ConfigItem(
+        keyName = "lootImageMinValue",
+        name = "Screenshot Min Value",
+        description = "The minimum combined loot value to send a screenshot. Must have 'Send Image' enabled",
+        position = 34,
+        section = lootSection
+    )
+    default int lootImageMinValue() {
+        return 0;
+    }
+
+    @ConfigItem(
         keyName = "lootIncludePlayer",
         name = "Include PK Loot",
         description = "Allow notifications for loot from player kills",
-        position = 34,
+        position = 35,
         section = lootSection
     )
     default boolean includePlayerLoot() {
@@ -448,7 +459,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "lootNotifMessage",
         name = "Notification Message",
         description = "The message to be sent through the webhook. Use %USERNAME% to insert your username, %LOOT% to insert the loot and %SOURCE% to show the source of the loot",
-        position = 35,
+        position = 36,
         section = lootSection
     )
     default String lootNotifyMessage() {
@@ -632,10 +643,21 @@ public interface DinkPluginConfig extends Config {
     }
 
     @ConfigItem(
+        keyName = "clueImageMinValue",
+        name = "Screenshot Min Value",
+        description = "The minimum combined value of the items to send a screenshot. Must have 'Send Image' enabled",
+        position = 74,
+        section = clueSection
+    )
+    default int clueImageMinValue() {
+        return 0;
+    }
+
+    @ConfigItem(
         keyName = "clueNotifMessage",
         name = "Notification Message",
         description = "The message to be sent through the webhook. Use %USERNAME% to insert your username, %CLUE% to insert the clue type, %LOOT% to show the loot obtained and %COUNT% to insert how many of those clue types you have completed",
-        position = 74,
+        position = 75,
         section = clueSection
     )
     default String clueNotifyMessage() {