+++ src/main/java/dinkplugin/DinkPlugin.java
@@ -48,7 +48,8 @@ import java.awt.Color;
 @PluginDescriptor(
     name = "Dink",
     description = "A notifier for sending webhooks to Discord or other custom destinations",
-    tags = { "loot", "logger", "collection", "pet", "death", "xp", "level", "notifications", "discord", "speedrun" }
+    tags = { "loot", "logger", "collection", "pet", "death", "xp", "level", "notifications", "discord", "speedrun",
+        "diary", "combat achievements", "combat task" }
 )
 public class DinkPlugin extends Plugin {