+++ src/main/java/dinkplugin/DinkPlugin.java
@@ -24,13 +24,17 @@ import net.runelite.api.events.StatChanged;
 import net.runelite.api.events.UsernameChanged;
 import net.runelite.api.events.VarbitChanged;
 import net.runelite.api.events.WidgetLoaded;
+import net.runelite.client.chat.ChatMessageManager;
+import net.runelite.client.chat.QueuedMessage;
 import net.runelite.client.config.ConfigManager;
 import net.runelite.client.eventbus.Subscribe;
+import net.runelite.client.events.ConfigChanged;
 import net.runelite.client.events.NpcLootReceived;
 import net.runelite.client.events.PlayerLootReceived;
 import net.runelite.client.plugins.Plugin;
 import net.runelite.client.plugins.PluginDescriptor;
 import net.runelite.client.plugins.loottracker.LootReceived;
+import net.runelite.client.util.ColorUtil;
 import net.runelite.client.util.Text;
 
 import javax.inject.Inject;
@@ -43,6 +47,10 @@ import javax.inject.Inject;
         "diary", "combat achievements", "combat task" }
 )
 public class DinkPlugin extends Plugin {
+
+    private @Inject ChatMessageManager chatManager;
+
+    private @Inject SettingsValidator validator;
 
     private @Inject CollectionNotifier collectionNotifier;
     private @Inject PetNotifier petNotifier;
@@ -76,6 +84,11 @@ public class DinkPlugin extends Plugin {
     @Provides
     DinkPluginConfig provideConfig(ConfigManager configManager) {
         return configManager.getConfig(DinkPluginConfig.class);
+    }
+
+    @Subscribe
+    public void onConfigChanged(ConfigChanged event) {
+        validator.onConfigChanged(event);
     }
 
     @Subscribe
@@ -149,6 +162,7 @@ public class DinkPlugin extends Plugin {
 
     @Subscribe
     public void onVarbitChanged(VarbitChanged event) {
+        validator.onVarbitChanged(event);
         diaryNotifier.onVarbitChanged(event);
     }
 
@@ -157,5 +171,20 @@ public class DinkPlugin extends Plugin {
         questNotifier.onWidgetLoaded(event);
         clueNotifier.onWidgetLoaded(event);
         speedrunNotifier.onWidgetLoaded(event);
+    }
+
+    public void addChatWarning(String message) {
+        String formatted = String.format("[%s] %s: %s",
+            ColorUtil.wrapWithColorTag(getName(), Utils.PINK),
+            "Warning",
+            ColorUtil.wrapWithColorTag(message, Utils.RED)
+        );
+
+        chatManager.queue(
+            QueuedMessage.builder()
+                .type(ChatMessageType.CONSOLE)
+                .runeLiteFormattedMessage(formatted)
+                .build()
+        );
     }
 }