+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Minor: Add level up setting to skip notifications below a specified level. (#97)
 - Minor: Retry failed webhook messages with exponential backoff. (#94)
+- Bugfix: Ensure boss name is included in boss slayer messages. (#88)
 - Dev: Bump mockito version to 4.10.0. (#95)
 
 ## 1.1.2