+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Minor: Add clue and loot notifier setting to skip screenshots for low item values. (#98)
 - Minor: Add level up setting to skip notifications below a specified level. (#97)
 - Minor: Retry failed webhook messages with exponential backoff. (#94)
 - Bugfix: Ensure boss name is included in boss slayer messages. (#88)