+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Unreleased
 
+- Minor: Add clue and loot notifier setting to skip screenshots for low item values. (#98)
+- Minor: Add level up setting to skip notifications below a specified level. (#97)
 - Minor: Retry failed webhook messages with exponential backoff. (#94)
 - Minor: Add warning logs for improper runescape settings that impact notifiers. (#92)
+- Bugfix: Ensure boss name is included in boss slayer messages. (#88)
 - Dev: Bump mockito version to 4.10.0. (#95)
 
 ## 1.1.2