+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Minor: Add level up setting to skip notifications below a specified level. (#97)
 - Minor: Retry failed webhook messages with exponential backoff. (#94)
 - Bugfix: Ensure boss name is included in boss slayer messages. (#88)
+- Dev: Add more plugin-hub search tags (#101)
 - Dev: Simplify notifier send image evaluation. (#102)
 - Dev: Bump mockito version to 4.10.0. (#95)