+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Minor: Add personal best time message to kill count notifier. (#106)
 - Minor: Add better descriptions for `Min Value` settings. (#100)
 - Minor: Add death notifier setting to disable kept item embeds. (#99)
 - Minor: Add clue and loot notifier setting to skip screenshots for low item values. (#98)